@@ -45,7 +45,7 @@
     "@nextcloud/axios": "^2.5.1",
     "@nextcloud/capabilities": "^1.2.0",
     "@nextcloud/cdav-library": "^2.1.1",
-    "@nextcloud/dialogs": "^5.3.8",
+    "@nextcloud/dialogs": "^7.0.0-rc.1",
     "@nextcloud/event-bus": "^3.3.2",
     "@nextcloud/files": "^3.12.0",
     "@nextcloud/initial-state": "^3.0.0",
@@ -55,7 +55,7 @@
     "@nextcloud/paths": "^2.2.1",
     "@nextcloud/router": "^3.0.1",
     "@nextcloud/sharing": "^0.2.5",
-    "@nextcloud/vue": "^8.27.0",
+    "@nextcloud/vue": "^9.0.0-rc.6",
     "@vueuse/core": "^11.3.0",
     "b64-to-blob": "^1.2.19",
     "camelcase": "^8.0.0",
@@ -68,19 +68,17 @@
     "mitt": "^3.0.1",
     "moment": "^2.30.1",
     "p-limit": "^7.0.0",
-    "pinia": "^2.3.1",
+    "pinia": "^3.0.3",
     "qreator": "^9.7.2",
     "string-natural-compare": "^3.0.1",
     "uuid": "^11.1.0",
     "validator": "^13.15.15",
-    "vue": "~2.7.16",
-    "vue-cropperjs": "^4.2.0",
+    "virtua": "^0.41.5",
+    "vue": "^3.5.18",
+    "vue-cropperjs": "^5.0.0",
     "vue-material-design-icons": "^5.3.1",
-    "vue-router": "^3.6.5",
-    "vue-virtual-scroll-list": "^2.3.5",
-    "vue-virtual-scroller": "^1.1.2",
-    "vuex": "^3.6.2",
-    "vuex-router-sync": "^5.0.0"
+    "vue-router": "^4.5.1",
+    "vuex": "^4.1.0"
   },
   "engines": {
     "node": "^20.0.0",
@@ -92,20 +90,18 @@
     "@nextcloud/eslint-config": "^8.4.2",
     "@nextcloud/stylelint-config": "^3.1.0",
     "@nextcloud/typings": "^1.9.1",
-    "@nextcloud/vite-config": "^1.5.6",
+    "@nextcloud/vite-config": "^2.4.0",
     "@playwright/test": "^1.54.1",
     "@types/jest": "^29.5.14",
-    "@vue/vue2-jest": "^29.2.6",
+    "@vue/vue3-jest": "^29.2.6",
     "babel-jest": "^29.7.0",
-    "eslint-import-resolver-typescript": "^3.7.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "jest-transform-stub": "^2.0.0",
     "ts-jest": "^29.4.0",
     "ts-loader": "^9.5.2",
     "typescript": "^5.8.3",
-    "vite": "^5.4.19",
-    "vue-template-compiler": "~2.7.16"
+    "vite": "^5.4.19"
   },
   "browserslist": [
     "extends @nextcloud/browserslist-config"
