@@ -6,8 +6,9 @@
 // eslint-disable-next-line import/no-unresolved, n/no-missing-import
 import 'vite/modulepreload-polyfill'
 
-import Vue from 'vue'
+import { createApp } from 'vue'
 import AdminSettings from './components/AdminSettings.vue'
+import LegacyGlobalMixin from './mixins/LegacyGlobalMixin.js'
 
 document.addEventListener('DOMContentLoaded', main)
 
@@ -15,9 +16,7 @@ document.addEventListener('DOMContentLoaded', main)
  *
  */
 function main() {
-	Vue.prototype.t = t
-
-	const View = Vue.extend(AdminSettings)
-	const view = new View()
-	view.$mount('#contacts-settings')
+	const app = createApp(AdminSettings)
+	app.mixin(LegacyGlobalMixin)
+	app.mount('#contacts-settings')
 }
