@@ -6,16 +6,12 @@
 // eslint-disable-next-line import/no-unresolved, n/no-missing-import
 import 'vite/modulepreload-polyfill'
 
-import { sync } from 'vuex-router-sync'
-import Vue from 'vue'
+import { createApp } from 'vue'
 
 import App from './ContactsRoot.vue'
 import router from './router/index.js'
 import store from './store/index.js'
-import logger from './services/logger.js'
-
-/** GLOBAL COMPONENTS AND DIRECTIVE */
-import { Tooltip as VTooltip } from '@nextcloud/vue'
+import LegacyGlobalMixin from './mixins/LegacyGlobalMixin.js'
 
 // Global scss sheets
 import './css/contacts.scss'
@@ -23,29 +19,17 @@ import './css/contacts.scss'
 // Dialogs css
 import '@nextcloud/dialogs/style.css'
 
-import { createPinia, PiniaVuePlugin } from 'pinia'
+import { createPinia } from 'pinia'
 
-Vue.use(PiniaVuePlugin)
+const app = createApp(App)
+
 const pinia = createPinia()
+app.use(pinia)
 
-// Register global directives
-Vue.directive('Tooltip', VTooltip)
+app.use(store)
+app.use(router)
 
-sync(store, router)
-
-Vue.prototype.t = t
-Vue.prototype.n = n
-
-Vue.prototype.appName = appName
-Vue.prototype.appVersion = appVersion
-Vue.prototype.logger = logger
-Vue.prototype.OC = window.OC
-Vue.prototype.OCA = window.OCA
-
-// enable devtools in development mode
-if (import.meta.env.MODE === 'development') {
-	Vue.config.devtools = true
-}
+app.mixin(LegacyGlobalMixin)
 
 // Force redirect if rewrite enabled but accessed through index.php
 if (window.location.pathname.split('/')[1] === 'index.php'
@@ -56,11 +40,4 @@ if (window.location.pathname.split('/')[1] === 'index.php'
 	})
 }
 
-export default new Vue({
-	el: '#content',
-	name: 'ContactsApp',
-	router,
-	store,
-	render: h => h(App),
-	pinia,
-})
+app.mount('#content')
