@@ -2,6 +2,7 @@
  * SPDX-FileCopyrightText: 2018 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
+
 module.exports = {
 	ignorePatterns: ['src/**/*.d.ts'],
 	globals: {
@@ -9,7 +10,7 @@ module.exports = {
 		appVersion: true,
 	},
 	extends: [
-		'@nextcloud/eslint-config/typescript',
+		'@nextcloud/eslint-config/vue3',
 	],
 	rules: {
 		// @nextcloud/vue has Button, Content components restricted by rule
