@@ -26,11 +26,11 @@ import {
 	NcAppContent as AppContent,
 	NcEmptyContent as EmptyContent,
 	NcLoadingIcon as IconLoading,
-	isMobile,
 } from '@nextcloud/vue'
 import AccountGroup from 'vue-material-design-icons/AccountGroupOutline.vue'
 import CircleDetails from '../CircleDetails.vue'
 import RouterMixin from '../../mixins/RouterMixin.js'
+import IsMobileMixin from '../../mixins/IsMobileMixin.ts'
 
 export default {
 	name: 'CircleContent',
@@ -43,7 +43,7 @@ export default {
 		IconLoading,
 	},
 
-	mixins: [isMobile, RouterMixin],
+	mixins: [IsMobileMixin, RouterMixin],
 
 	props: {
 		loading: {
