@@ -28,9 +28,8 @@
 					<IconContact :size="20" />
 				</template>
 				<template #counter>
-					<NcCounterBubble v-if="sortedContacts.length">
-						{{ sortedContacts.length }}
-					</NcCounterBubble>
+					<NcCounterBubble v-if="sortedContacts.length"
+						:count="sortedContacts.length" />
 				</template>
 			</AppNavigationItem>
 
@@ -56,9 +55,8 @@
 					<IconUser :size="20" />
 				</template>
 				<template #counter>
-					<NcCounterBubble v-if="ungroupedContacts.length">
-						{{ ungroupedContacts.length }}
-					</NcCounterBubble>
+					<NcCounterBubble v-if="ungroupedContacts.length"
+						:count="ungroupedContacts.length" />
 				</template>
 			</AppNavigationItem>
 
@@ -74,15 +72,14 @@
 					<IconRecentlyContacted :size="20" />
 				</template>
 				<template #counter>
-					<NcCounterBubble v-if="recentlyContactedContacts.contacts.length">
-						{{ recentlyContactedContacts.contacts.length }}
-					</NcCounterBubble>
+					<NcCounterBubble v-if="recentlyContactedContacts.contacts.length"
+						:count="recentlyContactedContacts.contacts.length" />
 				</template>
 			</AppNavigationItem>
 
 			<AppNavigationCaption id="newgroup"
+				v-model:menu-open="isNewGroupMenuOpen"
 				:force-menu="true"
-				:menu-open.sync="isNewGroupMenuOpen"
 				:name="t('contacts', 'Contact groups')"
 				@click.prevent.stop="toggleNewGroupMenu">
 				<template #actionsTriggerIcon>
@@ -160,11 +157,13 @@
 				<AppNavigationItem :aria-label="t('contacts', 'Open the contacts app settings')"
 					:name="CONTACTS_SETTINGS"
 					@click="showContactsSettings">
-					<Cog slot="icon" :size="20" />
+					<template #icon>
+						<Cog :size="20" />
+					</template>
 				</AppNavigationItem>
 			</div>
 		</template>
-		<ContactsSettings :open.sync="showSettings" />
+		<ContactsSettings v-model:open="showSettings" />
 	</AppNavigation>
 </template>
 
