@@ -9,31 +9,31 @@
 				name: 'contact',
 				params: {
 					selectedGroup: selectedChart,
-					selectedContact: data.key,
+					selectedContact: chartData.key,
 				},
 			}">
 				<Avatar :disable-tooltip="true"
-					:display-name="data.fullName"
+					:display-name="chartData.fullName"
 					:is-no-user="true"
 					:size="60"
-					:url="data.photoUrl"
+					:url="chartData.photoUrl"
 					class="org-chart-node__avatar" />
 			</router-link>
 			<div class="panel" />
 			<div class="main-container">
 				<h3 class="fullName">
-					{{ data.fullName }}
+					{{ chartData.fullName }}
 				</h3>
 				<h4 class="title">
-					{{ data.title }}
+					{{ chartData.title }}
 				</h4>
 			</div>
 			<div class="description">
-				<p v-if="data._directSubordinates">
-					{{ t('contacts', 'Manages') }}: {{ data._directSubordinates }}
+				<p v-if="chartData._directSubordinates">
+					{{ t('contacts', 'Manages') }}: {{ chartData._directSubordinates }}
 				</p>
-				<p v-if="data._totalSubordinates">
-					{{ t('contacts', 'Oversees') }}: {{ data._totalSubordinates }}
+				<p v-if="chartData._totalSubordinates">
+					{{ t('contacts', 'Oversees') }}: {{ chartData._totalSubordinates }}
 				</p>
 			</div>
 		</div>
@@ -49,7 +49,7 @@ export default {
 		Avatar,
 	},
 	props: {
-		data: {
+		chartData: {
 			type: Object,
 			default: () => {},
 		},
