@@ -63,13 +63,15 @@
 				</EmptyContent>
 
 				<!-- Searched & picked entities -->
-				<VirtualList v-else-if="searchSet.length > 0 && availableEntities.length > 0"
+				<VList v-else-if="searchSet.length > 0 && availableEntities.length > 0"
+					v-slot="{ item }"
 					class="entity-picker__options"
-					data-key="id"
-					:data-sources="availableEntities"
-					:data-component="EntitySearchResult"
-					:estimate-size="44"
-					:extra-props="{ selection: selectionSet, onClick }" />
+					:data="availableEntities">
+					<EntitySearchResult :key="item.id"
+						:source="item"
+						:selection="selectionSet"
+						:on-click="onClick" />
+				</VList>
 
 				<EmptyContent v-else-if="searchQuery" :name="t('contacts', 'No results')">
 					<template #icon>
@@ -96,7 +98,7 @@
 
 <script>
 import debounce from 'debounce'
-import VirtualList from 'vue-virtual-scroll-list'
+import { VList } from 'virtua/vue'
 import {
 	NcButton as Button,
 	NcEmptyContent as EmptyContent,
@@ -120,7 +122,8 @@ export default {
 		IconSearch,
 		IconLoading,
 		Modal,
-		VirtualList,
+		VList,
+		EntitySearchResult,
 	},
 
 	props: {
@@ -207,7 +210,6 @@ export default {
 			canInviteGuests: !!window?.OCA?.Guests?.openGuestDialog,
 			searchQuery: '',
 			localSelection: {},
-			EntitySearchResult,
 		}
 	},
 
@@ -352,7 +354,7 @@ export default {
 		 * @param {object} entity the entity to remove
 		 */
 		onDelete(entity) {
-			this.$delete(this.selectionSet, entity.id, entity)
+			delete this.selectionSet[entity.id]
 			console.debug('Removing entity from selection', entity)
 		},
 
@@ -363,12 +365,12 @@ export default {
 		 */
 		onClick(entity) {
 			if (entity.id in this.selectionSet) {
-				this.$delete(this.selectionSet, entity.id)
+				delete this.selectionSet[entity.id]
 				console.debug('Removed entity to selection', entity)
 				return
 			}
 
-			this.$set(this.selectionSet, entity.id, entity)
+			this.selectionSet[entity.id] = entity
 			console.debug('Added entity to selection', entity)
 		},
 
