@@ -3,8 +3,7 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-import Vue from 'vue'
-import Vuex, { Store } from 'vuex'
+import { Store } from 'vuex'
 
 import addressbooks from './addressbooks.js'
 import circles from './circles.js'
@@ -13,8 +12,6 @@ import groups from './groups.js'
 import importState from './importState.js'
 
 import isCirclesEnabled from '../services/isCirclesEnabled.js'
-
-Vue.use(Vuex)
 
 const mutations = {}
 
