@@ -2,8 +2,8 @@
  * SPDX-FileCopyrightText: 2022 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
+
 import Contact from '../../../src/models/contact'
-import { Property } from 'ical.js'
 
 const getPropertyLines = (property, vcard) => {
 	return vcard.match(new RegExp(`^${property}[;:].*`, 'gmi'))
@@ -25,31 +25,28 @@ describe('Test stripping quotes from TYPE', () => {
 		property = contact.vCard.addPropertyWithValue('TEl', '+00 123 456 789')
 	})
 
-	test('Test stripping quotes from SINGLE TYPE', (done) => {
+	test('Test stripping quotes from SINGLE TYPE', () => {
 		property.setParameter('type', ['VOICE'])
 		const line = getPropertyLines('TEL', contact.toStringStripQuotes())[0]
 
 		expect(line).toStrictEqual('TEL;TYPE=VOICE:+00 123 456 789')
-		done()
 	})
 
-	test('Test stripping quotes from MULTIPLE TYPES', (done) => {
+	test('Test stripping quotes from MULTIPLE TYPES', () => {
 		property.setParameter('type', ['WORK', 'VOICE'])
 		const line = getPropertyLines('TEL', contact.toStringStripQuotes())[0]
 
 		expect(line).toStrictEqual('TEL;TYPE=WORK,VOICE:+00 123 456 789')
-		done()
 	})
 
-	test('Test stripping quotes from MULTIPLE SPLIT TYPES', (done) => {
+	test('Test stripping quotes from MULTIPLE SPLIT TYPES', () => {
 		property.setParameter('type', ['WORK,VOICE'])
 		const line = getPropertyLines('TEL', contact.toStringStripQuotes())[0]
 
 		expect(line).toStrictEqual('TEL;TYPE=WORK,VOICE:+00 123 456 789')
-		done()
 	})
 
-	test('Test stripping quotes from MULTIPLE SPLIT TYPES and MULTIPLE PROPERTIES', (done) => {
+	test('Test stripping quotes from MULTIPLE SPLIT TYPES and MULTIPLE PROPERTIES', () => {
 		const property2 = contact.vCard.addPropertyWithValue('TEl', '+99 876 543 210')
 		property.setParameter('type', ['WORK,VOICE'])
 		property2.setParameter('type', ['HOME'])
@@ -60,7 +57,6 @@ describe('Test stripping quotes from TYPE', () => {
 			'TEL;TYPE=WORK,VOICE:+00 123 456 789',
 			'TEL;TYPE=HOME:+99 876 543 210',
 		])
-		done()
 	})
 
 })
